backend/requirements.txt
Werkzeug==2.2.2
Django==4.0.5
numpy==1.22.4
pandas==2.0.3
psycopg2-binary==2.9.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
zappa==0.54.1
pillow==9.1.1
social-auth-app-django==5.0.0
google-api-python-client==2.127.0
google-auth==2.29.0
openai==1.23.6